reorganised plotting module
