+++ src/pymodaq/daq_logger.py
@@ -8,17 +8,13 @@ Contains all objects related to the DAQ_Scan module, to do automated scans, savi
 
 import sys
 from collections import OrderedDict
-import numpy as np
-import logging
 
 import pymodaq.daq_utils.parameter.ioxml
 
-from pyqtgraph.parametertree import Parameter, ParameterTree
-from qtpy import QtGui, QtWidgets
-from qtpy.QtCore import QObject, Slot, QThread, Signal, QLocale, Qt
+from qtpy import QtWidgets
+from qtpy.QtCore import QObject, Slot, QThread, Signal, Qt
 
-from pymodaq.daq_utils.managers.modules_manager import ModulesManager
-from pymodaq.daq_utils.plotting.qled import QLED
+from pymodaq.daq_utils.plotting.widgets.qled import QLED
 from pymodaq.daq_utils import daq_utils as utils
 from pymodaq.daq_utils import gui_utils as gutils
 from pymodaq.daq_utils.h5modules import H5Logger
