--- src/pymodaq/daq_utils/plotting/image_viewer_proj/__init__.py
@@ -1,1 +1,0 @@
-from .image_viewer import Image_Viewer
