+++ src/pymodaq/daq_utils/plotting/items/crosshair.py
