+++ src/pymodaq/daq_utils/plotting/widgets/qled.py
